Updated Navbar.js
Navigation bar was updated. The side toolbar was updated to be a temporary Drawer instead of a permanent one.
The top Navbar has Deaths, Vaccinations, and Total as buttons now, as well as having a menu icon that should minimize and expand it, but the menu icon currently does not do anything.